modules/locationWorker/requirements.txt
kafka-python
sqlalchemy
geoalchemy2
shapely
psycopg2